+++ src/index.ts
@@ -41,6 +41,22 @@ import { set } from './set';
 
 import { isPointer } from './isPointer';
 
+// ### encodeToken(string: string) -> string
+
+// encodes a string for safe use as a json pointer token.
+
+// * [documentation](encodeToken.html)
+
+import { encodeToken } from './encodeToken';
+
+// ### decodeToken(string: string) -> string
+
+// decodes a json token unsecaping special character.
+
+// * [documentation](decodeToken.html)
+
+import { decodeToken } from './decodeToken';
+
 // ## Usage
 
 // Install JST via npm.
@@ -60,5 +76,7 @@ export {
   dereference,
   get,
   isPointer,
+  encodeToken,
+  decodeToken,
   set,
 };